vdab-daily-sync/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
tenacity>=8.2.3
APScheduler>=3.10.4
python-dateutil>=2.8.2
